Table-based lists, menu/settings polish, search-DB + configuration fixes
UI:
- Replace one-per-row bordered cards with dense Mantine Tables via composable
  ItemTable/ItemRow primitives (replacing ItemCard); rows keep the right-click
  context menu and kebab.
- Swap the hand-rolled accordion for Mantine Accordion (interactive headers) on
  the documents page, and turn the document back card into an interactive
  breadcrumb button.
- Center the configuration loading spinner so it no longer overlaps the favorite
  button in the insert menu.
- Soften settings buttons (default/light variants instead of solid fills) and
  restore context-menu intent colors (show/hide-all, admin submenu).

Backend:
- Fix empty configuration parameters: the parameters column is already json-mode,
  so drop the redundant JSON.stringify/parse that left the stored array empty.
- Build the search index on the backend and persist it to a new
  libraries.search_db D1 column: rebuilt after each document load and lazily on
  first /search-db request, removing the manual frontend push.
