+++ src/frontend/search/search-results.tsx
@@ -1,6 +1,7 @@
 import { ReactNode } from "react";
 import { Position, SearchFilters, SearchHit, doSearch } from "./search";
 import { InsertableCard } from "../cards/insertable-card";
+import { ItemTable } from "../cards/card-components";
 import { SectionError, SectionLoading } from "../common/app-zero-state";
 import { NoSearchResultError, SearchCallout } from "./search-errors";
 import { useLibraryQuery, useSearchDbQuery } from "../queries";
@@ -64,7 +65,7 @@ export function SearchResults(props: SearchResultsProps): ReactNode {
     return (
         <>
             {callout}
-            {resultCards}
+            <ItemTable>{resultCards}</ItemTable>
         </>
     );
 }
